Made r2d2_sim.m compatible with Matlab.
Evidently, Octave and Matlab slightly differ in how they handle cell arrays.  I had a 2x2 cell array for the platoon directives and indexed it with numbers 1, 2, 3, 4 in Octave and that worked fine.  But Matlab identifies them as 1,3,2,4.  So I changed it to a single row cell array, and now both programs should identify them the same way.